add telegram notification and create telegram bot w telegraf

+++ src/modules/cron/cron.service.ts
@@ -3,13 +3,15 @@ import { Injectable } from "@nestjs/common";
 import { MailService } from "../libs/mail/mail.service";
 import { Cron } from "@nestjs/schedule";
 import { StorageService } from "../libs/storage/storage.service";
+import { TelegramService } from "../libs/telegram/telegram.service";
 
 @Injectable()
 export class CronService {
   constructor(
     private readonly prismaService: PrismaService,
     private readonly mailService: MailService,
-    private readonly storageService: StorageService
+    private readonly storageService: StorageService,
+    private readonly telegramService: TelegramService
   ) {}
 
   // @Cron("*/10 * * * * *")
@@ -25,11 +27,24 @@ export class CronService {
           lte: sevenDaysAgo,
         },
       },
+      include: {
+        notificationSettings: true,
+        streams: true,
+      },
     });
     for (const user of deactivatedAccounts) {
       await this.mailService.sendAccountDeletion(user.email);
 
-      this.storageService.remove(user.avatar);
+      if (user.notificationSettings.telegramNotifications && user.telegramId) {
+        await this.telegramService.sendAccountDeletion(user.telegramId);
+      }
+
+      if (user.avatar) {
+        this.storageService.remove(user.avatar);
+      }
+      if (user.streams.thumbnailUrl) {
+        this.storageService.remove(user.streams.thumbnailUrl);
+      }
     }
 
     await this.prismaService.user.deleteMany({
